Nomikos_Final/requirements.txt
streamlit
langchain
langchain-community
langchain-groq
langchain-pinecone
langchain-huggingface
python-dotenv
pypdf
pinecone-client
sentence-transformers
sqlalchemy
psycopg2-binary